backend/requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
sqlalchemy==1.4.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==0.19.0
alembic==1.7.4
dataclasses-json==0.5.7
email-validator==2.1.0.post1
typing-extensions>=4.8.0
httpx>=0.24.0
